lot s of fixes

+++ Assets/ArenaGames/Scripts/Helpers/AGTimeController.cs
@@ -5,6 +5,6 @@ namespace ArenaGames
     public static class AGTimeController
     {
         public static long TimestampMS => ((DateTimeOffset) DateTime.Now).ToUnixTimeMilliseconds();
-        public static long Timestamp => ((DateTimeOffset) DateTime.Now).ToUnixTimeSeconds();
+        public static int Timestamp => (int) ((DateTimeOffset) DateTime.Now).ToUnixTimeSeconds();
     }
 }
